# caresoft_core/caresoft_core/Connected Services/CoreWebApi/OpenAPI.nswag.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "caresoft_core",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Consultorio/get/{id}": {
      "get": {
        "tags": [
          "Consultorio"
        ],
        "operationId": "get4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Consultorio"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consultorio"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consultorio"
                }
              }
            }
          }
        }
      }
    },
    "/api/Consultorio/add": {
      "post": {
        "tags": [
          "Consultorio"
        ],
        "operationId": "add5",
        "parameters": [
          {
            "name": "IdConsultorio",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Nombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Direccion",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Telefono",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Consulta",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Consultum"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Consultorio/update": {
      "put": {
        "tags": [
          "Consultorio"
        ],
        "operationId": "update5",
        "parameters": [
          {
            "name": "IdConsultorio",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Nombre",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Direccion",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Telefono",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Consulta",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/Consultum"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Consultorio/delete/{id}": {
      "delete": {
        "tags": [
          "Consultorio"
        ],
        "operationId": "delete5",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Afeccion": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idAfeccion": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "consultaCodigos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Consultum"
            }
          },
          "idIngresos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Ingreso"
            }
          }
        }
      },
      "Aseguradora": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idAseguradora": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "correo": {
            "type": "string",
            "nullable": true
          },
          "autorizacions": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Autorizacion"
            }
          }
        }
      },
      "Autorizacion": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idAutorizacion": {
            "type": "integer",
            "format": "int32"
          },
          "idAseguradora": {
            "type": "integer",
            "format": "int32"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "montoAsegurado": {
            "type": "number",
            "format": "double"
          },
          "consulta": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Consultum"
            }
          },
          "facturaProducto": {
            "$ref": "#/components/schemas/FacturaProducto"
          },
          "facturaServicio": {
            "$ref": "#/components/schemas/FacturaServicio"
          },
          "idAseguradoraNavigation": {
            "$ref": "#/components/schemas/Aseguradora"
          },
          "ingresos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Ingreso"
            }
          }
        }
      },
...
      "Consultorio": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idConsultorio": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "consulta": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Consultum"
            }
          }
        }
      },
      "Consultum": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "consultaCodigo": {
            "type": "string",
            "nullable": true
          },
          "documentoPaciente": {
            "type": "string",
            "nullable": true
          },
          "documentoMedico": {
            "type": "string",
            "nullable": true
          },
          "idConsultorio": {
            "type": "integer",
            "format": "int32"
          },
          "idAutorizacion": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "motivo": {
            "type": "string",
            "nullable": true
          },
          "comentarios": {
            "type": "string",
            "nullable": true
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "costo": {
            "type": "number",
            "format": "double"
          },
          "documentoMedicoNavigation": {
            "$ref": "#/components/schemas/PerfilUsuario"
          },
          "documentoPacienteNavigation": {
            "$ref": "#/components/schemas/PerfilUsuario"
          },
          "factura": {
            "$ref": "#/components/schemas/Factura"
          },
          "idAutorizacionNavigation": {
            "$ref": "#/components/schemas/Autorizacion"
          },
          "idConsultorioNavigation": {
            "$ref": "#/components/schemas/Consultorio"
          },
          "prescripcionProductos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/PrescripcionProducto"
            }
          },
          "idAfeccions": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Afeccion"
            }
          },
          "servicioCodigos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Servicio"
            }
          }
        }
      },
      "Cuentum": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idCuenta": {
            "type": "integer",
            "format": "int32"
          },
          "documentoUsuario": {
            "type": "string",
            "nullable": true
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "documentoUsuarioNavigation": {
            "$ref": "#/components/schemas/PerfilUsuario"
          },
          "facturas": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Factura"
            }
          },
          "pagos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Pago"
            }
          }
        }
      },
...
      "Factura": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "facturaCodigo": {
            "type": "string",
            "nullable": true
          },
          "idCuenta": {
            "type": "integer",
            "format": "int32"
          },
          "consultaCodigo": {
            "type": "string",
            "nullable": true
          },
          "idIngreso": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "idSucursal": {
            "type": "integer",
            "format": "int32"
          },
          "documentoCajero": {
            "type": "string",
            "nullable": true
          },
          "montoSubtotal": {
            "type": "number",
            "format": "double"
          },
          "montoTotal": {
            "type": "number",
            "format": "double"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "consultaCodigoNavigation": {
            "$ref": "#/components/schemas/Consultum"
          },
          "documentoCajeroNavigation": {
            "$ref": "#/components/schemas/PerfilUsuario"
          },
          "facturaProductos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FacturaProducto"
            }
          },
          "facturaServicios": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FacturaServicio"
            }
          },
          "idCuentaNavigation": {
            "$ref": "#/components/schemas/Cuentum"
          },
          "idIngresoNavigation": {
            "$ref": "#/components/schemas/Ingreso"
          },
          "idSucursalNavigation": {
            "$ref": "#/components/schemas/Sucursal"
          },
          "idMetodoPagos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/MetodoPago"
            }
          }
        }
      },
...
      "FacturaProducto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "facturaCodigo": {
            "type": "string",
            "nullable": true
          },
          "idProducto": {
            "type": "integer",
            "format": "int32"
          },
          "idAutorizacion": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "resultados": {
            "type": "string",
            "nullable": true
          },
          "costo": {
            "type": "number",
            "format": "double"
          },
          "facturaCodigoNavigation": {
            "$ref": "#/components/schemas/Factura"
          },
          "idAutorizacionNavigation": {
            "$ref": "#/components/schemas/Autorizacion"
          },
          "idProductoNavigation": {
            "$ref": "#/components/schemas/Producto"
          }
        }
      },
...
      "FacturaServicio": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "facturaCodigo": {
            "type": "string",
            "nullable": true
          },
          "servicioCodigo": {
            "type": "string",
            "nullable": true
          },
          "idAutorizacion": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "resultados": {
            "type": "string",
            "nullable": true
          },
          "costo": {
            "type": "number",
            "format": "double"
          },
          "facturaCodigoNavigation": {
            "$ref": "#/components/schemas/Factura"
          },
          "idAutorizacionNavigation": {
            "$ref": "#/components/schemas/Autorizacion"
          },
          "servicioCodigoNavigation": {
            "$ref": "#/components/schemas/Servicio"
          }
        }
      },
      "Ingreso": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idIngreso": {
            "type": "integer",
            "format": "int32"
          },
          "documentoPaciente": {
            "type": "string",
            "nullable": true
          },
          "documentoEnfermero": {
            "type": "string",
            "nullable": true
          },
          "documentoMedico": {
            "type": "string",
            "nullable": true
          },
          "consultaCodigo": {
            "type": "string",
            "nullable": true
          },
          "idAutorizacion": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "numSala": {
            "type": "integer",
            "format": "int32"
          },
          "costoEstancia": {
            "type": "number",
            "format": "double"
          },
          "fechaIngreso": {
            "type": "string",
            "format": "date-time"
          },
          "fechaAlta": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "documentoEnfermeroNavigation": {
            "$ref": "#/components/schemas/PerfilUsuario"
          },
          "documentoMedicoNavigation": {
            "$ref": "#/components/schemas/PerfilUsuario"
          },
          "documentoPacienteNavigation": {
            "$ref": "#/components/schemas/PerfilUsuario"
          },
          "factura": {
            "$ref": "#/components/schemas/Factura"
          },
          "idAutorizacionNavigation": {
            "$ref": "#/components/schemas/Autorizacion"
          },
          "numSalaNavigation": {
            "$ref": "#/components/schemas/Sala"
          },
          "idAfeccions": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Afeccion"
            }
          }
        }
      },
...
      "MetodoPago": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idMetodoPago": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "facturaCodigos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Factura"
            }
          }
        }
      },
...
      "Pago": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idPago": {
            "type": "integer",
            "format": "int32"
          },
          "idCuenta": {
            "type": "integer",
            "format": "int32"
          },
          "montoPagado": {
            "type": "number",
            "format": "double"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "idCuentaNavigation": {
            "$ref": "#/components/schemas/Cuentum"
          }
        }
      },
      "PerfilUsuario": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "documento": {
            "type": "string",
            "nullable": true
          },
          "tipoDocumento": {
            "type": "string",
            "nullable": true
          },
          "numLicenciaMedica": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "apellido": {
            "type": "string",
            "nullable": true
          },
          "genero": {
            "type": "string",
            "nullable": true
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date-time"
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "correo": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "rol": {
            "type": "string",
            "nullable": true
          },
          "consultumDocumentoMedicoNavigations": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Consultum"
            }
          },
          "consultumDocumentoPacienteNavigations": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Consultum"
            }
          },
          "cuentum": {
            "$ref": "#/components/schemas/Cuentum"
          },
          "facturas": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Factura"
            }
          },
          "ingresoDocumentoEnfermeroNavigations": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Ingreso"
            }
          },
          "ingresoDocumentoMedicoNavigations": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Ingreso"
            }
          },
          "ingresoDocumentoPacienteNavigations": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Ingreso"
            }
          },
          "reservaServicioDocumentoMedicoNavigations": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ReservaServicio"
            }
          },
          "reservaServicioDocumentoPacienteNavigations": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ReservaServicio"
            }
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          }
        }
      },
      "PrescripcionProducto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "consultaCodigo": {
            "type": "string",
            "nullable": true
          },
          "idProducto": {
            "type": "integer",
            "format": "int32"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "consultaCodigoNavigation": {
            "$ref": "#/components/schemas/Consultum"
          },
          "idProductoNavigation": {
            "$ref": "#/components/schemas/Producto"
          }
        }
      },
      "Producto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idProducto": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "costo": {
            "type": "number",
            "format": "double"
          },
          "loteDisponible": {
            "type": "integer",
            "format": "int32"
          },
          "facturaProductos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FacturaProducto"
            }
          },
          "prescripcionProductos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/PrescripcionProducto"
            }
          },
          "rncProveedors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Proveedor"
            }
          }
        }
      },
...
      "Proveedor": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "rncProveedor": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "correo": {
            "type": "string",
            "nullable": true
          },
          "idProductos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Producto"
            }
          }
        }
      },
...
      "ReservaServicio": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idReserva": {
            "type": "integer",
            "format": "int32"
          },
          "documentoPaciente": {
            "type": "string",
            "nullable": true
          },
          "documentoMedico": {
            "type": "string",
            "nullable": true
          },
          "servicioCodigo": {
            "type": "string",
            "nullable": true
          },
          "fechaReservada": {
            "type": "string",
            "format": "date-time"
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "documentoMedicoNavigation": {
            "$ref": "#/components/schemas/PerfilUsuario"
          },
          "documentoPacienteNavigation": {
            "$ref": "#/components/schemas/PerfilUsuario"
          },
          "servicioCodigoNavigation": {
            "$ref": "#/components/schemas/Servicio"
          }
        }
      },
...
      "Sala": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "numSala": {
            "type": "integer",
            "format": "int32"
          },
          "estado": {
            "type": "string",
            "nullable": true
          },
          "ingresos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Ingreso"
            }
          }
        }
      },
...
      "Servicio": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "servicioCodigo": {
            "type": "string",
            "nullable": true
          },
          "idTipoServicio": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "descripcion": {
            "type": "string",
            "nullable": true
          },
          "costo": {
            "type": "number",
            "format": "double"
          },
          "facturaServicios": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FacturaServicio"
            }
          },
          "idTipoServicioNavigation": {
            "$ref": "#/components/schemas/TipoServicio"
          },
          "reservaServicios": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ReservaServicio"
            }
          },
          "consultaCodigos": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Consultum"
            }
          }
        }
      },
...
      "Sucursal": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idSucursal": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "direccion": {
            "type": "string",
            "nullable": true
          },
          "telefono": {
            "type": "string",
            "nullable": true
          },
          "facturas": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Factura"
            }
          }
        }
      },
...
      "TipoServicio": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idTipoServicio": {
            "type": "integer",
            "format": "int32"
          },
          "nombre": {
            "type": "string",
            "nullable": true
          },
          "servicios": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Servicio"
            }
          }
        }
      },
...
      "Usuario": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "usuarioCodigo": {
            "type": "string",
            "nullable": true
          },
          "documentoUsuario": {
            "type": "string",
            "nullable": true
          },
          "usuarioContra": {
            "type": "string",
            "nullable": true
          },
          "documentoUsuarioNavigation": {
            "$ref": "#/components/schemas/PerfilUsuario"
          }
        }
      },
...
    }
  }
}